Move files around, add schema.json
