+++ client/package.json
@@ -15,7 +15,6 @@
     "@types/react-dom": "^16.9.6",
     "apollo": "^2.27.0",
     "apollo-boost": "^0.4.7",
-    "graphql": "^15.0.0",
     "graphql-tag": "^2.10.3",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
